open/requirements.txt
openai
python-dotenv
pydantic
loguru
